clean とりあえず現状の責務で分けられそうなところをファイル分割
